Support receiving samples from HA Prometheus pairs.
Samples for a given cluster are differentiated by the 'cluster' label (by default).
The pairs are differentiated by a '__replica__' label (by default).
The replica label is removed in the distributor to prevent mass series churn on failover.

We track the "elected" replica per tenant & cluster in another KVStore.
This can optionally be the same KVStore are the existing ring.
Replicas are "reelected" after a timeout.
Writes to the KVStore are rate limited by tenant/cluster with a 202, which is not logged.
Sent samples for the non-elected master are silently dropped.

Also:
- Refactor ring/kvstore so that we pass in codec and add WatchPrefix.
- Move KVClient Mock from ring config to KVStore config.
- Have KVStore config take an optional prefix so we can reuse it's register flags function and still keep the existing ring.consul flags.

Signed-off-by: Callum Styan <callumstyan@gmail.com>
Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>