Big change: Remove the label copying in QueryStream.
This change removes the use of LabelPair and wire.Bytes.  We directly marshal and unmarshal from the bytes on the wire into labels.Labels, via LabelsAdapater.

I've migrated everywhere I could find to these new types.  We can almost completely remove wire.Bytes, execpt for its use in the caching index proto.  So I moved the type there.

I'd draw attention to the use of yoloString and the use of unsafe for casting between []LabelAdapater and labels.Labels.  They should be safe as we take a copy of the label string in the inverted index.

Also:
- changed all occurances of `__name__` to model.MetricNameLabel.
- removed a bunch of unused functions for converting between old types.
- make the FromXXXToYYY naming consistent in compat.go.
- fork stdlib fnv32 hashing like we have with fnv64.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>