Refactor: Unify metadata and samples sends to ingesters (#2410)
* Refactor: Unify metadata and samples sends to ingesters

Previously, we had samples and metadata using the ring on their own to
send data to the ingesters. Since the `WriteRequest` supports metadata
alongside samples there's no reason to split the calls so we unify them. There's an
additional tiny refactor to inline the metadata validation as it is just
a single function call.

Signed-off-by: gotjosh <josue@grafana.com>

* Address review comments

Signed-off-by: gotjosh <josue@grafana.com>